+++ tests/vector-test.c
@@ -35,9 +35,9 @@ int cgd_main(int argc, char **argv)
     struct timeval start = { 0 }, stop = { 0 };
     gettimeofday(&start, NULL);
 
-    s_log_debug("Running %u iterations...", N_ITERATIONS);
+    s_log_info("Running %u iterations...", N_ITERATIONS);
     for (u32 i = 0; i < N_ITERATIONS; i++) {
-        s_log_info("Iteration %u", i);
+        s_log_debug("Iteration %u", i);
         if (test()) {
             s_log_info("Test result is FAIL");
             return EXIT_FAILURE;
@@ -76,7 +76,7 @@ static i32 test(void)
     struct large_struct *vector_large = NULL;
     u64 *vector_cloned = NULL;
 
-    s_log_debug("Initializing vectors...");
+    s_log_trace("Initializing vectors...");
     vector_small = vector_new(u64);
     if (vector_small == NULL)
         goto_error("Failed to initialize vector_small!");
@@ -85,11 +85,11 @@ static i32 test(void)
     if (vector_large == NULL)
         goto_error("Failed to initialize vector_large!");
 
-    s_log_debug("Testing vector_reserve...");
+    s_log_trace("Testing vector_reserve...");
     vector_reserve(&vector_small, N_SMALL_ITEMS);
     vector_reserve(&vector_large, 0);
 
-    s_log_debug("Testing vector_push_back...");
+    s_log_trace("Testing vector_push_back...");
 
     for (u32 i = 0; i < N_SMALL_ITEMS; i++)
         vector_push_back(&vector_small, small_items[i]);
@@ -113,7 +113,7 @@ static i32 test(void)
         goto_error("`vector_large->items` does not match `large_items`");
 
     u32 vsmall_n_popped_back = random_u32() % ((N_SMALL_ITEMS / 2) - 1) + 1;
-    s_log_debug("Testing vector_pop_back (and vector_back) with %u items...",
+    s_log_trace("Testing vector_pop_back (and vector_back) with %u items...",
         vsmall_n_popped_back);
     for (u32 i = 0; i < vsmall_n_popped_back; i++)
         vector_pop_back(&vector_small);
@@ -127,7 +127,7 @@ static i32 test(void)
         goto_error("`vector_small->items` does not match `small_items`");
 
     u32 vlarge_insert_index = random_u32() % (N_LARGE_STRUCTS - 1);
-    s_log_debug("Testing vector_insert (index %u, max %u) and vector_erase...",
+    s_log_trace("Testing vector_insert (index %u, max %u) and vector_erase...",
         vlarge_insert_index, N_LARGE_STRUCTS - 1);
     vector_insert(&vector_large, vlarge_insert_index,
         (struct large_struct){ .arr = { 0 }, .str = "inserted item" }
@@ -142,7 +142,7 @@ static i32 test(void)
             vector_large[vlarge_insert_index].str);
 
     u64 *vsmall_begin_val = vector_begin(vector_small);
-    s_log_debug("Testing vector_begin, vector_front and vector_end...",
+    s_log_trace("Testing vector_begin, vector_front and vector_end...",
         vsmall_begin_val
     );
     if (vsmall_begin_val != vector_small)
@@ -155,14 +155,14 @@ static i32 test(void)
 
     u64 *vsmall_end_val_p = vector_end(vector_small);
     u64 vsmall_pushed_val = (u64)random_u32();
-    s_log_debug("vector_end() returned %p -> %lu, pushed value is %lu",
+    s_log_trace("vector_end() returned %p -> %lu, pushed value is %lu",
         vsmall_end_val_p, *vsmall_end_val_p, vsmall_pushed_val);
     vector_push_back(&vector_small, vsmall_pushed_val);
     if (vector_back(vector_small) != *vsmall_end_val_p || *vsmall_end_val_p != vsmall_pushed_val)
         goto_error("vector_end test failed, expected value %lu, got %lu",
             vsmall_pushed_val, *vsmall_end_val_p);
 
-    s_log_debug("Testing vector_clone, vector_empty and vector_clear...");
+    s_log_trace("Testing vector_clone, vector_empty and vector_clear...");
     vector_cloned = vector_clone(vector_small);
     if (
         vector_cloned == NULL ||
@@ -176,14 +176,14 @@ static i32 test(void)
     if (vector_empty(vector_small) || !vector_empty(vector_cloned))
         goto_error("vector_empty test failed");
 
-    s_log_debug("Testing vector_capacity and vector_shrink_to_fit...");
+    s_log_trace("Testing vector_capacity and vector_shrink_to_fit...");
     vector_shrink_to_fit(&vector_cloned);
     if (vector_capacity(vector_cloned) != VECTOR_MINIMUM_CAPACITY__)
         goto_error("vector_shrink_to_fit test failed; expected value %u, got %u",
             VECTOR_MINIMUM_CAPACITY__, vector_capacity(vector_cloned));
 
     u32 vsmall_new_size = random_u32() % (N_SMALL_ITEMS - vsmall_n_popped_back - 1);
-    s_log_debug("Testing vector_resize -> new_size %u and vector_at...",
+    s_log_trace("Testing vector_resize -> new_size %u and vector_at...",
         vsmall_new_size);
     vector_resize(&vector_small, vsmall_new_size);
     if (vsmall_new_size > 0) {
@@ -191,7 +191,7 @@ static i32 test(void)
             goto_error("vector_resize test failed; expected value %lu, got %lu",
                 vector_at(vector_small, vsmall_new_size), vector_back(vector_small));
     } else {
-        s_log_debug("Testing vector_push_back when size is 0");
+        s_log_trace("Testing vector_push_back when size is 0");
         vsmall_pushed_val = random_u32();
         vector_push_back(&vector_small, vsmall_pushed_val);
         if (vector_size(vector_small) != 1) {